# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/add_in_basket": {
            "post": {
                "tags": [
                    "Basket"
                ],
                "summary": "add_in_basket",
                "description": "add_in_basket",
                "operationId": "add_in_basket",
                "requestBody": {
                    "description": "Pass user register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "product_id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "register created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/get_my_basket": {
            "post": {
                "tags": [
                    "Basket"
                ],
                "summary": "get_my_basket",
                "description": "get_my_basket",
                "operationId": "get_my_basket",
                "requestBody": {
                    "description": "Pass user register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "register created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/delete_all_basket": {
            "post": {
                "tags": [
                    "Basket"
                ],
                "summary": "delete_all_basket",
                "description": "delete_all_basket",
                "operationId": "delete_all_basket",
                "requestBody": {
                    "description": "Pass user register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "register created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/minus_basket_product": {
            "post": {
                "tags": [
                    "Basket"
                ],
                "summary": "minus_basket_product",
                "description": "minus_basket_product",
                "operationId": "minus_basket_product",
                "requestBody": {
                    "description": "Pass user register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "product_id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "register created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/get_shops": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "get_shops",
                "description": "get_shops",
                "operationId": "get_shops",
                "requestBody": {
                    "description": "Auth Token Required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "register created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/get_category": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "get_category",
                "description": "get_category",
                "operationId": "get_category",
                "requestBody": {
                    "description": "Auth Token Required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "register created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/get_all_made_in": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "get_all_made_in",
                "description": "get_all_made_in",
                "operationId": "get_all_made_in",
                "requestBody": {
                    "description": "Auth Token Required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "register created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/get_all_taste": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "get_all_taste",
                "description": "get_all_taste",
                "operationId": "get_all_taste",
                "requestBody": {
                    "description": "Auth Token Required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "register created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}